+++ packages/wrangler/src/deployment-bundle/bindings.ts
