+++ packages/wrangler/src/versions/upload.ts
@@ -4,6 +4,7 @@ import { blue, gray } from "@cloudflare/cli/colors";
 import { syncAssets } from "../assets";
 import { fetchResult } from "../cfetch";
 import { configFileName, formatConfigSnippet, printBindings } from "../config";
+import { getBindings } from "../deployment-bundle/bindings";
 import { bundleWorker } from "../deployment-bundle/bundle";
 import {
 	printBundleSize,
@@ -21,7 +22,6 @@ import {
 	getWrangler1xLegacyModuleReferences,
 } from "../deployment-bundle/module-collection";
 import { validateNodeCompatMode } from "../deployment-bundle/node-compat";
-import { getBindings } from "../deployment-bundle/provision";
 import { loadSourceMaps } from "../deployment-bundle/source-maps";
 import { confirm } from "../dialogs";
 import { getMigrationsToUpload } from "../durable";
